+++ acoular/aiaa/__init__.py
@@ -1,0 +1,12 @@
+# ------------------------------------------------------------------------------
+# Copyright (c) Acoular Development Team.
+# ------------------------------------------------------------------------------
+"""Provides classes for importing AIAA Array Benchmarks.
+
+.. autosummary::
+    :toctree: generated/
+
+    aiaa
+"""
+
+from .aiaa import CsmAIAABenchmark, MicAIAABenchmark, TimeSamplesAIAABenchmark, TriggerAIAABenchmark
