+++ tests/unittests/test_aiaa.py
@@ -14,8 +14,8 @@ def test_load_mics(aiaa_bechmark_time_data_file):
 
     """
     with pytest.deprecated_call():
-        mic_geom = ac.tools.MicAIAABenchmark(from_file=aiaa_bechmark_time_data_file)
-    mic_geom = ac.tools.MicAIAABenchmark(file=aiaa_bechmark_time_data_file)
+        mic_geom = ac.aiaa.MicAIAABenchmark(from_file=aiaa_bechmark_time_data_file)
+    mic_geom = ac.aiaa.MicAIAABenchmark(file=aiaa_bechmark_time_data_file)
     assert mic_geom.num_mics == 8
     assert mic_geom.pos.shape == (3, 8)
     assert mic_geom.pos_total.shape == (3, 8)
@@ -32,8 +32,8 @@ def test_load_csm(aiaa_bechmark_csm_file):
         Path to the AIAA benchmark file with cross spectral matrix data (fixture
     """
     with pytest.deprecated_call():
-        csm = ac.tools.CsmAIAABenchmark(name=aiaa_bechmark_csm_file)
-    csm = ac.tools.CsmAIAABenchmark(file=aiaa_bechmark_csm_file)
+        csm = ac.aiaa.CsmAIAABenchmark(name=aiaa_bechmark_csm_file)
+    csm = ac.aiaa.CsmAIAABenchmark(file=aiaa_bechmark_csm_file)
     assert csm.num_channels == 8
     assert csm.csm.shape == (csm.fftfreq().shape[0], csm.num_channels, csm.num_channels)
 
@@ -49,8 +49,8 @@ def test_load_time(aiaa_bechmark_time_data_file):
         Path to the AIAA benchmark file with time data (fixture).
     """
     with pytest.deprecated_call():
-        time_data = ac.tools.TimeSamplesAIAABenchmark(name=aiaa_bechmark_time_data_file)
-    time_data = ac.tools.TimeSamplesAIAABenchmark(file=aiaa_bechmark_time_data_file)
+        time_data = ac.aiaa.TimeSamplesAIAABenchmark(name=aiaa_bechmark_time_data_file)
+    time_data = ac.aiaa.TimeSamplesAIAABenchmark(file=aiaa_bechmark_time_data_file)
     assert time_data.num_channels == 8
     assert time_data.num_samples == 480
     assert time_data.sample_freq == 48000
@@ -69,8 +69,8 @@ def test_load_trigger(aiaa_bechmark_trigger_file):
         Path to the AIAA benchmark file with trigger data (fixture).
     """
     with pytest.deprecated_call():
-        trigger_data = ac.tools.TriggerAIAABenchmark(name=aiaa_bechmark_trigger_file)
-    trigger_data = ac.tools.TriggerAIAABenchmark(file=aiaa_bechmark_trigger_file)
+        trigger_data = ac.aiaa.TriggerAIAABenchmark(name=aiaa_bechmark_trigger_file)
+    trigger_data = ac.aiaa.TriggerAIAABenchmark(file=aiaa_bechmark_trigger_file)
     assert trigger_data.num_channels == 1
     assert trigger_data.num_samples == 480
     assert trigger_data.sample_freq == 48000
